+++ test/SensorSimulator/sensor.cpp
@@ -1,4 +1,7 @@
-#include "Sensor.h"
+#include "sensor.h"
+
+namespace sensor_simulator
+{
 
 Sensor::Sensor(Ekf* ekf)
 {
@@ -26,3 +29,5 @@ bool Sensor::is_time_to_send(uint32_t time)
 {
 	return (time >= _time_last_data_sent) && ((time - _time_last_data_sent) >= _update_period);
 }
+
+} // namespace sensor_simulator
